Some more clean-up to Chatter.
  * Fixed a bug with positioning on drag.
  * Moved TerminalPatch out.
  * Still not great code but it'll do for now.
